Add custom node name support for cluster nodes
Introduces the ability to set a custom node name for cluster identification via the config file, defaulting to the hostname if not specified. Updates Hazelcast cluster configuration to use this node name as a member attribute and instance name. Also updates the JSON schema to document the new NodeName property.